LayoutTests/http/tests/security/contentSecurityPolicy/script-src-blocked-error-event-expected.txt

CONSOLE MESSAGE: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/script-set-value.js' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".


PASS Script element has error on bad script-src 
